Updated script

+++ Scripts/armadev.ps1
@@ -1,35 +1,58 @@
 
 $armaRptPath="$env:USERPROFILE\AppData\Local\Arma 3"
 $program="$($MyInvocation.MyCommand.Name.Replace('.ps1', ''))"
+
 function help() {
-    Write-Host "$program
+    Write-Host "$program"
+        "`nUsage:"
+        "    $program rpt [-c] [-w]"
+        "    $program p [--set <path>] [--mount] [--umount]"
+        "    $program tools"
+        "    $program (-h | --help)"
 
-Usage:
-    $program rpt [-c] [-w]
-    $program p [--set <path>] [--mount] [--umount]
-    $program (-h | --help)
+        "`nCommands:"
+        "    rpt     RPT log handeling"
+        "    p       P-Drive handeling"
+        "    tools   Show avalible Arma 3 Tools"
 
-Commands:
-    rpt     RPT log handeling
-    p       P-Drive handeling
+        "`nOptions:"
+        "    rpt -c, --clear                 clear all arma 3 rpt logs"
+        "    rpt -w, --watch < | filter>     watch latest created RPT log"
+        "    p --set <PATH>                  set the path of the P-Drive"
+        "                                    note: You can set this to any path allowing"
+        "                                    multiply P-Drive locations"
+        "    p --mount                       mount the P-Drive using the set path"
+        "    p --umount                      unmount the P-Drive"
+        "    -h, --help                      show this help"
+    exit 0
+}
 
-Options:
-    rpt -c, --clear                 clear all arma 3 rpt logs
-    rpt -w, --watch < | filter>     watch latest created RPT log
-    p --set <PATH>                  set the path of the P-Drive
-                                    note: You can set this to any path allowing
-                                    multiply P-Drive locations
-    p --mount                       mount the P-Drive using the set path
-    p --umount                      unmount the P-Drive
-    -h, --help                      show this help"
+function tools() {
+    function Test-Tools() { Get-Command $args[0] -errorAction SilentlyContinue }
+
+    Write-Host "Discoverd tools:"
+    $regpath="Registry::HKEY_CURRENT_USER\SOFTWARE\Bohemia Interactive\Arma 3 Tools"
+    if ( Get-ItemPropertyValue $regpath -Name path -errorAction SilentlyContinue ) { Write-Host " > Arma 3 Tools $(Get-ItemPropertyValue $regpath -Name version -errorAction SilentlyContinue)"}
+    if ( Test-Tools armake ) { Write-Host " > armake $(armake --version)"}
+    if ( Test-Tools armake2 ) { Write-Host " > armake2 $(armake2 --version)"}
+    if ( Test-Tools hemtt ) { Write-Host " > $(hemtt --version)"}
+    $regpath="Registry::HKEY_CURRENT_USER\SOFTWARE\Mikero"
+    $mikeroTools=$(Get-ChildItem $regpath -Name)
+    ForEach ( $tool in $mikeroTools ) {
+        Write-Host " > Mikero $tool"
+    }
     exit 0
 }
 
 function rpt() {
+    #Check avalible logs
+    $logs = @(Get-ChildItem -Path "$armaRptPath\*" -Include *.rpt)
+    
     if ( $args[0] -eq '--clear' -or $args[0] -eq '-c' ) {
-        Write-Host "Clearing RPT logs"
+        if ( $logs.length -eq 0 ) { Write-Host "No RPT logs to clear"; exit 0 }
+        Write-Host "Clearing RPT logs..." -confirm
         Get-ChildItem -Path "$armaRptPath\*" -Include *.rpt | Remove-Item
-        Write-Host "Done"
+        Write-Host "All rpt logs have been removed."
         exit 0
     }
 
@@ -39,19 +62,20 @@ function rpt() {
             Write-Host "No logfiles are avalible to watch"
             exit 1
         }
-        Write-Host "Starting watch of latest RPT log file"
-        tail $latestLog $args[0]
+        Write-Host "Starting watch of latest RPT log file: $($latestLog.name)"
+        if ( $latestLog -and -not $args[0] ) {
+            Get-Content $latestLog -wait -tail $($Host.UI.RawUI.WindowSize.Height-2)
+        } else {
+            Get-Content $latestLog -wait -tail $($Host.UI.RawUI.WindowSize.Height-2) | Select-String -NoEmphasis -Patter $args[0]
+        }
         exit 0
     }
 
     Set-Location -Path "$armaRptPath"
-    $logs = @(Get-ChildItem -Path "$armaRptPath\*" -Include *.rpt)
-    if ( $logs.length -eq 0 ) {
-        Write-Host "There are currently no RPT logs"
-        exit 0
-    }
-    Write-Host "Avalible RPT logs:"
-    Write-Host "$logs.name"
+    if ( $logs.length -eq 0 ) { Write-Host "There are currently no RPT logs" -ForegroundColor DarkGray; exit 0 }
+    Write-Host "Avalible RPT logs ($($logs.count)):"
+    $logs.name | ForEach-Object { Write-Host " > $_" -ForegroundColor DarkGray }
+
     exit 0
 }
 
@@ -72,27 +96,30 @@ function p() {
     if ( -not $(Test-Path "$pdrivePath") ) { Write-Host "Given path '$pdrivePath' is not valid"; exit 1 }
 
     # mount and unmount
+    $regpath="Registry::HKEY_CURRENT_USER\SOFTWARE\Bohemia Interactive\Arma 3 Tools"
+    $mountToolPath=if ( Get-ItemPropertyValue $regpath -Name path -errorAction SilentlyContinue ) { $(Get-ItemPropertyValue $regpath -Name path -errorAction SilentlyContinue) }
+    $mountTool= Join-Path $mountToolPath -ChildPath "WorkDrive\WorkDrive.exe"
     if ( $args[0] -eq '--mount' ) {
-        #& "C:\Program Files (x86)\Steam\steamapps\common\Arma 3 Tools\WorkDrive\WorkDrive.exe" /Mount $pdrivePath
-        Start-Process -NoNewWindow -FilePath "C:\Program Files (x86)\Steam\steamapps\common\Arma 3 Tools\WorkDrive\WorkDrive.exe" -ArgumentList "/Mount", "P" , $pdrivePath
+        if ( Test-Path -Path "P:" ) { Write-Host "P-Drive already mounted"; exit 1 }
+        Start-Process -FilePath $mountTool -ArgumentList "/Mount", "P" , $pdrivePath
         exit 0
     }
     if ( $args[0] -eq '--umount' ) {
-        Start-Process -NoNewWindow -FilePath "C:\Program Files (x86)\Steam\steamapps\common\Arma 3 Tools\WorkDrive\WorkDrive.exe" -ArgumentList "/Dismount"
+        if ( -not $(Test-Path -Path "P:") ) { Write-Host "P-Drive not mounted"; exit 1 }
+        Start-Process -FilePath $mountTool -ArgumentList "/Dismount"
         exit 0
     }
 
     # defult behaviour
-    if ( Test-Path -Path "P:" ) {
-        Set-Location -Path $pdrivePath
-        exit 0
-    } else {
-        Write-Host "Arma 3 P: drive is not mounted pleace mount it and try again."
-        exit 1
+    if ( -not $(Test-Path -Path "P:") ) {
+        Write-Host "Warning: Arma 3 P-drive is not mounted" -ForegroundColor Yellow
     }
+    Set-Location -Path $pdrivePath
+    exit 0
 }
 
 if ( $args[0] -eq '--help' -or $args[0] -eq '-h' ) { help }
+if ( $args[0] -eq 'tools' ) { tools $args[1] }
 if ( $args[0] -eq 'rpt' ) { rpt $args[1] }
 if ( $args[0] -eq 'p' ) { p $args[1] $args[2] }
 
@@ -100,10 +127,10 @@ if ( $args ) {
     Write-Host "'$args' is not a supported parameter"
     exit 1
 } else {
-    Write-Host "$program
-    
-Usage:
-    armadev rpt [-c] [-w]
-    armadev p [--set <path>] [--mount] [--umount]
-    armadev (-h | --help)"
+    Write-Host "$program"
+        "`nUsage:"
+        "    $program rpt [-c] [-w]"
+        "    $program p [--set <path>] [--mount] [--umount]"
+        "    $program tools"
+        "    $program (-h | --help)"
 }
